# app/api/admin_faiss.py
"""
Admin API endpoints for FAISS index management
"""
from fastapi import APIRouter, HTTPException, Depends
from datetime import datetime, timezone
import uuid

from app.schemas.faiss_management import (
    IndexBuildRequest,
    IndexBuildResponse,
    IndexHotSwapRequest,
    IndexHotSwapResponse,
    IndexStatusResponse,
    PerformanceMetrics,
    MigrationStatus
)
from app.dependencies import require_admin
from app.tasks import celery_app

router = APIRouter(prefix="/admin/index", tags=["admin-faiss"])

# Global index metadata (in production, use Redis)
current_index_metadata = {
    "version": "v1",
    "type": "IndexFlatIP",
    "vector_count": 100000,
    "dimensions": 384,
    "memory_size_mb": 150.5,
    "build_time_seconds": 12.5,
    "performance_metrics": {
        "p95_search_latency_ms": 25.0,
        "p99_search_latency_ms": 45.0,
        "memory_usage_percent": 35.0,
        "search_throughput_qps": 2000.0
    },
    "migration_status": {
        "is_migrating": False,
        "current_version": "v1",
        "target_version": None,
        "traffic_percentage": 100.0,
        "migration_phase": "none"
    },
    "migration_thresholds": {
        "vector_count_trigger": 500000,
        "latency_trigger_ms": 100,
        "memory_trigger_percent": 80
    }
}


@router.post("/build", response_model=IndexBuildResponse)
async def build_index(
    request: IndexBuildRequest,
    admin: dict = Depends(require_admin)
):
    """
    Trigger FAISS index build (admin only).
    
    Builds new FAISS index from embedding batches with configurable
    optimization options and validation.
    
    Args:
        request: Index build parameters
        admin: Authenticated admin user
    
    Returns:
        Build task details
    
    Example:
        POST /admin/index/build
        {
            "embedding_batch_paths": ["s3://bucket/embeddings/batch1", ...],
            "index_type": "IndexIVFFlat",
            "optimization_options": {
                "normalize_vectors": true,
                "memory_optimization": false,
                "validation_enabled": true
            }
        }
    """
    # Generate new version
    new_version = f"v{int(current_index_metadata['version'][1:]) + 1}"
    
    # Trigger Celery task
    task = celery_app.send_task(
        'build_faiss_index',
        args=[
            request.embedding_batch_paths,
            request.index_type,
            new_version
        ],
        kwargs={
            'normalize_vectors': request.optimization_options.normalize_vectors,
            'memory_optimization': request.optimization_options.memory_optimization,
            'validation_enabled': request.optimization_options.validation_enabled
        },
        queue='index_build'
    )
    
    # Estimate build time (15 seconds for 10K vectors as baseline)
    estimated_vectors = len(request.embedding_batch_paths) * 1000
    estimated_time = (estimated_vectors / 10000) * 15
    
    return IndexBuildResponse(
        task_id=str(task.id),
        index_version=new_version,
        estimated_build_time_seconds=estimated_time,
        status_endpoint=f"/admin/index/status?version={new_version}"
    )


@router.post("/hot-swap", response_model=IndexHotSwapResponse)
async def hot_swap_index(
    request: IndexHotSwapRequest,
    admin: dict = Depends(require_admin)
):
    """
    Perform zero-downtime index hot-swap (admin only).
    
    Implements safe index deployment with:
    - Pre-deployment validation
    - Performance benchmark comparison
    - Gradual traffic migration
    - Health check validation
    - Automatic rollback on regression
    
    Args:
        request: Hot-swap parameters
        admin: Authenticated admin user
    
    Returns:
        Hot-swap operation status
    
    Example:
        POST /admin/index/hot-swap
        {
            "new_index_version": "v2",
            "gradual_migration": true,
            "rollback_on_regression": true,
            "latency_regression_threshold": 10.0
        }
    """
    swap_id = f"swap_{uuid.uuid4()}"
    
    # Pre-deployment validation
    validation_results = {
        "dimension_check": True,
        "integrity_check": True,
        "memory_check": True,
        "performance_benchmark": True
    }
    
    # Update migration status
    current_index_metadata["migration_status"] = {
        "is_migrating": True,
        "current_version": current_index_metadata["version"],
        "target_version": request.new_index_version,
        "traffic_percentage": 0.0,
        "migration_phase": "validation"
    }
    
    # Trigger Celery task for hot-swap
    task = celery_app.send_task(
        'hot_swap_index',
        args=[request.new_index_version, swap_id],
        kwargs={
            'gradual_migration': request.gradual_migration,
            'rollback_on_regression': request.rollback_on_regression,
            'latency_threshold': request.latency_regression_threshold
        },
        queue='index_swap'
    )
    
    return IndexHotSwapResponse(
        swap_id=swap_id,
        status="initiated",
        current_phase="validation",
        validation_results=validation_results,
        estimated_completion_seconds=60.0 if request.gradual_migration else 10.0
    )


@router.get("/status", response_model=IndexStatusResponse)
async def get_index_status(
    admin: dict = Depends(require_admin),
    version: str = None
):
    """
    Get current index status and metadata (admin only).
    
    Returns comprehensive index information including performance
    metrics, migration status, and trigger thresholds.
    
    Args:
        admin: Authenticated admin user
        version: Optional specific version to query
    
    Returns:
        Index status and metadata
    
    Example:
        GET /admin/index/status
        
        Response:
            {
                "version": "v1",
                "type": "IndexFlatIP",
                "vector_count": 100000,
                "dimensions": 384,
                "memory_size_mb": 150.5,
                "build_time_seconds": 12.5,
                "performance_metrics": {...},
                "migration_status": {...},
                "migration_thresholds": {...}
            }
    """
    # In production, query specific version from storage
    if version and version != current_index_metadata["version"]:
        raise HTTPException(
            status_code=404,
            detail=f"Index version '{version}' not found"
        )
    
    return IndexStatusResponse(
        version=current_index_metadata["version"],
        type=current_index_metadata["type"],
        vector_count=current_index_metadata["vector_count"],
        dimensions=current_index_metadata["dimensions"],
        memory_size_mb=current_index_metadata["memory_size_mb"],
        build_time_seconds=current_index_metadata["build_time_seconds"],
        performance_metrics=PerformanceMetrics(**current_index_metadata["performance_metrics"]),
        migration_status=MigrationStatus(**current_index_metadata["migration_status"]),
        migration_thresholds=current_index_metadata["migration_thresholds"]
    )