# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: APFA - Agentic Personalized Financial Advisor API
  version: 1.0.0
  description: |
    Production-ready API for AI-powered financial advisory system with multi-agent architecture,
    RAG (Retrieval-Augmented Generation), and background job processing.
    
    ## Features
    - JWT-based authentication
    - Real-time task monitoring via WebSocket
    - Background job processing with Celery
    - FAISS vector search for document retrieval
    - LLM-powered advice generation
    
    ## Rate Limits
    - Global: 10 requests/minute per IP
    - Authentication: 10 login attempts/minute per IP
    - Advice generation: 10 requests/minute per authenticated user
    
  contact:
    name: APFA Backend Team
    email: apfa-backend@company.com
  license:
    name: Proprietary
    url: https://apfa.io/license
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://staging-api.apfa.io
    description: Staging environment
  - url: https://api.apfa.io
    description: Production environment
...
paths:
...
  /api/admin/index:
    get:
      tags:
        - Admin - Index
      summary: Get FAISS indexes
      description: Get list of all FAISS indexes with metadata
      operationId: getIndexes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Index list retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FAISSIndex'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /token endpoint
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key (deprecated, use JWT)
  schemas:
...
    FAISSIndex:
      type: object
      required:
        - version
        - vectorCount
        - dimensions
        - indexType
        - isActive
      properties:
        version:
          type: string
          description: SHA256 hash (first 8 chars)
          example: abc12345
        vectorCount:
          type: integer
          minimum: 0
          example: 100000
        dimensions:
          type: integer
          example: 384
        indexType:
          type: string
          enum: [IndexFlatIP, IndexIVFFlat, IndexIVFPQ]
          example: IndexFlatIP
        memoryMB:
          type: number
          format: float
          description: Memory usage in MB
          example: 153.6
        createdAt:
          type: string
          format: date-time
          example: 2025-10-11T12:00:00Z
        isActive:
          type: boolean
          description: Whether this is the currently active index
          example: true
        buildDuration:
          type: number
          format: float
          description: Time to build index in seconds
          nullable: true
          example: 5.2
        migrationUrgency:
          type: number
          format: float
          minimum: 0
          maximum: 100
          description: |
            Migration urgency score (0-100)
            < 70: No action
            70-90: Plan migration
            > 90: Migrate immediately
          example: 45.2
...
security:
  - bearerAuth: []